# Faster blockdiag for uniform input value-index types