@irhum 2020-05-24
added another method to deal with the empty input case, `blockdiag()`, otherwise it defaults to 
```julia
blockdiag(X::AbstractSparseMatrixCSC{Tv, Ti}...) where {Tv, Ti <: Integer}
```

instead of 
```julia
blockdiag(X::AbstractSparseMatrixCSC...)
```

resulting in a `Tv not defined` error, making an explicit definition necessary. 

The pull request, I would suggest is important given the case where all inputs have the same types for the value and index would be more common (and would benefit from increased performance) than the heterogenous case