@dkarrasch 2020-05-24
You'll need to rename the function at the call sites as well.