@irhum 2020-05-24
The function call sites have been changed to make the appropriate `_blockdiag` call.

Since we also need to account for the case where no inputs are given, a clean solution seems to be to pass it as an optional argument to one of the methods. Since we're explicitly handing it, `isempty(X)` in

```julia
Ti = isempty(X) ? Int : promote_type(map(x->eltype(rowvals(x)), X)...)
```

is always false, and hence can be replaced with

```
Ti = promote_type(map(x->eltype(rowvals(x)), X)...)
```

Opinions?